Move all ordering logic into libs/jsonschema
