+++ libs/jsonschema/schema.go
@@ -4,17 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"os"
-	"slices"
-	"strings"
 )
-
-// Property defines a single property of a struct schema.
-// This type is not used in the schema itself but rather to
-// return the pair of a property name and its schema.
-type Property struct {
-	Name   string
-	Schema *Schema
-}
 
 // defines schema for a json object
 type Schema struct {
@@ -83,30 +73,6 @@ func (schema *Schema) validate() error {
 		}
 	}
 	return nil
-}
-
-// PropertiesOrdered returns the items of the `prope`
-func (s *Schema) PropertiesOrdered() []Property {
-	order := make(map[string]int)
-	out := make([]Property, len(s.Properties))
-	for key, property := range s.Properties {
-		order[key] = property.Order
-		out = append(out, Property{
-			Name:   key,
-			Schema: property,
-		})
-	}
-
-	// Sort the properties by order and then by name.
-	slices.SortFunc(out, func(a, b Property) int {
-		k := order[a.Name] - order[b.Name]
-		if k != 0 {
-			return k
-		}
-		return strings.Compare(a.Name, b.Name)
-	})
-
-	return out
 }
 
 func Load(path string) (*Schema, error) {
